feat: Add functionality to view who the user is following
Implements a new section to display a list of users that the current promoter is following. This involves adding a new service function to fetch this data from Firestore and integrating it into the UI.

Also includes a minor refactor in `ErrorBoundary.tsx` to change `class ErrorBoundary extends Component` to `class ErrorBoundary extends React.Component`.