feat: Update promoter group status directly
Introduces a new service function `confirmPromoterGroupEntry` to directly update a promoter's `hasJoinedGroup` status in Firestore. This bypasses the need for an Admin Cloud Function, streamlining the process on the client-side and improving performance.

The `StatusCheck` page is updated to utilize this new function when a promoter accepts the rules and has not yet joined the group.

Additionally, minor refactoring in `ErrorBoundary` component to use React class component standard syntax.